+++ tests/TestCase.php
@@ -27,6 +27,7 @@ class TestCase extends Orchestra
     {
         $driver = getenv('DB_CONNECTION') ?: 'sqlite';
         $app['config']->set('app.key', 'base64:'.base64_encode(str_repeat('a', 32)));
+        $app['config']->set('cache.default', 'array');
         $app['config']->set('database.default', 'testing');
         $app['config']->set(
             'database.connections.testing',